nl.cljs
Hi, apart from a few minor alterations (for example, changing "your" into "you" when "you" was to be translated... I've actuallly added quite a lot. I noticed the English file had a lot of extra lines which aren't in the Dutch translation. I've added these, and translated them too. Hope it will be of value. The new document is one line longer than the English one because of one additional line in the Dutch which I did not dare remove; line 227, "other". Line 63 I've left untranslated, because I believe it has to remain as is. Hope I could be of help.